api/requirements.txt
Flask
Flask-Cors
boto3
pypdf
python-docx
gunicorn